move CollectDemographicFields inline
